pyproject.toml
[tool.poetry]
name = "datagovharvester"
version = "0.1.0"
description = ""
authors = ["Jin-Sun-tts <jin.sun@gsa.gov>"]
readme = "README.md"
packages = [{include = "datagovharvester"}]

[tool.poetry.dependencies]
python = ">=3.9"
jsonschema = "^4.17.3"
psycopg2 = "^2.9.6"
redis = "^4.5.5"
celery = "^5.2.7"
requests = "^2.30.0"
python-dotenv = "^1.0.0"
pytest-celery = "^0.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.0"
ruff = "^0.0.261"
pytest-cov = "^4.0.0"

[tool.ruff]
# Enable pycodestyle (`E`) and Pyflakes (`F`) codes by default.
select = ["E", "F"]
ignore = []

# Allow autofix for all enabled rules (when `--fix`) is provided.
fixable = ["A", "B", "C", "D", "E", "F", "G", "I", "N", "Q", "S", "T", "W", "ANN", "ARG", "BLE", "COM", "DJ", "DTZ", "EM", "ERA", "EXE", "FBT", "ICN", "INP", "ISC", "NPY", "PD", "PGH", "PIE", "PL", "PT", "PTH", "PYI", "RET", "RSE", "RUF", "SIM", "SLF", "TCH", "TID", "TRY", "UP", "YTT"]
unfixable = []

# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]

# Same as Black.
line-length = 88

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

# Assume Python 3.10.
target-version = "py310"

[tool.ruff.mccabe]
# Unlike Flake8, default to a complexity level of 10.
max-complexity = 10


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
